requirements.txt
scikit-learn==1.4.0
keras==2.15.0
nltk==3.8.1
Flask==3.0.1
Flask-Cors==4.0.0
tensorflow==2.15.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
hazm==0.10.0
mysql-connector-python==8.3.0